refactor: Create PaperItem to mutualize code from PaperLine
